Rename mock_timer.h to fake_timer.h (to match class name
